move GoogleCredentials to credentials.py
This reduces the size of core.py
